add resnet.py and update
